added integration tests

+++ internal/service/team_service/team.go
@@ -1,6 +1,7 @@
 package team_service
 
 import (
+	"fmt"
 	"github.com/labstack/echo/v4"
 	"team-management/common/echo_ctx"
 	"team-management/internal/models/primary/team"
@@ -17,6 +18,9 @@ func NewTeamService(teamRepo *team_repo.TeamRepo) *TeamService {
 	}
 }
 func (t *TeamService) CreateTeam(ctx echo.Context, req CreateTeamRequest) (*team.Team, error) {
+	if req.Name == "" {
+		return nil, fmt.Errorf("invalid name")
+	}
 	createTeamRequest := team_repo.CreateTeamRequest{
 		Name:      req.Name,
 		CreatedBy: echo_ctx.GetUserId(ctx),
@@ -28,6 +32,9 @@ func (t *TeamService) CreateTeam(ctx echo.Context, req CreateTeamRequest) (*team
 	return teamObj, nil
 }
 func (t *TeamService) AddMember(ctx echo.Context, req MemberRequest) error {
+	if req.UserId == "" {
+		return fmt.Errorf("invalid name")
+	}
 	request := team_repo.MemberRequest{
 		UserId: req.UserId,
 		TeamId: req.TeamId,
@@ -42,6 +49,13 @@ func (t *TeamService) RemoveMember(ctx echo.Context, req MemberRequest) error {
 		UserId: req.UserId,
 		TeamId: req.TeamId,
 	}
+	memberExists, err := t.teamRepo.MemberExists(ctx, request)
+	if err != nil {
+		return err
+	}
+	if !memberExists {
+		return fmt.Errorf("member does not exist")
+	}
 	if err := t.teamRepo.RemoveMember(ctx, request); err != nil {
 		return err
 	}
@@ -51,6 +65,13 @@ func (t *TeamService) MakeAdmin(ctx echo.Context, req MemberRequest) error {
 	request := team_repo.MemberRequest{
 		UserId: req.UserId,
 		TeamId: req.TeamId,
+	}
+	memberExists, err := t.teamRepo.MemberExists(ctx, request)
+	if err != nil {
+		return err
+	}
+	if !memberExists {
+		return fmt.Errorf("member does not exist")
 	}
 	if err := t.teamRepo.MakeAdmin(ctx, request); err != nil {
 		return err
